Scrum 75 쿠폰 결제 추가

+++ src/main/java/com/sparta/projectblue/domain/coupon/dto/CreateCouponRequestDto.java
@@ -30,7 +30,7 @@ public class CreateCouponRequestDto {
 
     @NotNull private int currentQuantity;
 
-    @NotNull private int discountValue;
+    @NotNull private Long discountValue;
 
     @NotNull private LocalDateTime startDate;
